# internal/api/docs/swagger.yaml
basePath: /v1
definitions:
...
  admin_product_handler.CategoryResponseDto:
    properties:
      description:
        example: EVENT
        type: string
      id:
        example: 3
        type: integer
      name:
        example: event
        type: string
    type: object
  admin_product_handler.CreateInstallmentsRequestDto:
    properties:
      fee:
        type: number
      id:
        example: 1
        type: integer
      paymentTypeId:
        type: integer
      tariff:
        example: 2.3
        type: number
    required:
    - id
    - paymentTypeId
    type: object
  admin_product_handler.CreateInstallmentsResponseDto:
    properties:
      fee:
        example: 1.5
        type: number
      id:
        example: 1
        type: integer
      installmentId:
        example: 1
        type: integer
      paymentTypeId:
        example: 1
        type: integer
      tariff:
        example: 2.3
        type: number
    type: object
  admin_product_handler.CreateRequestDto:
    properties:
      active:
        example: true
        type: boolean
      categoryId:
        example: 3
        minimum: 1
        type: integer
      description:
        example: Fresh and fiery red hot chilly peppers, perfect for adding a spicy
          kick to your dishes.
        maxLength: 2000
        type: string
      discountPrice:
        example: 4.99
        type: number
      image:
        example: https://example.com/images/red-hot-chilly-peppers.jpg
        maxLength: 2000
        type: string
      imageMobile:
        example: https://example.com/images/red-hot-chilly-peppers-mobile.jpg
        maxLength: 2000
        type: string
      imageThumbnail:
        example: https://example.com/images/red-hot-chilly-peppers-thumbnail.jpg
        maxLength: 2000
        type: string
      installments:
        items:
          $ref: '#/definitions/admin_product_handler.CreateInstallmentsRequestDto'
        minItems: 1
        type: array
      name:
        example: Red Hot Chilly Peppers
        maxLength: 255
        type: string
      price:
        example: 5.99
        type: number
      stock:
        $ref: '#/definitions/admin_product_handler.CreateStockRequestDto'
    required:
    - categoryId
    - image
    - imageMobile
    - imageThumbnail
    - installments
    - name
    - price
    type: object
  admin_product_handler.CreateResponseDto:
    properties:
      active:
        example: true
        type: boolean
      categoryId:
        example: 3
        type: integer
      createdAt:
        example: "2023-01-01T00:00:00Z"
        type: string
      createdBy:
        example: 1
        type: integer
      description:
        example: Fresh and fiery red hot chilly peppers, perfect for adding a spicy
          kick to your dishes.
        type: string
      discountPrice:
        example: 4.99
        type: number
      id:
        example: 1
        type: integer
      image:
        example: https://example.com/images/red-hot-chilly-peppers.jpg
        type: string
      imageMobile:
        example: https://example.com/images/red-hot-chilly-peppers-mobile.jpg
        type: string
      imageThumbnail:
        example: https://example.com/images/red-hot-chilly-peppers-thumbnail.jpg
        type: string
      installments:
        items:
          $ref: '#/definitions/admin_product_handler.CreateInstallmentsResponseDto'
        type: array
      isDeleted:
        example: false
        type: boolean
      name:
        example: Red Hot Chilly Peppers
        type: string
      price:
        example: 5.99
        type: number
      stock:
        $ref: '#/definitions/admin_product_handler.CreateStockResponseDto'
      updatedAt:
        example: "2023-01-01T00:00:00Z"
        type: string
      updatedBy:
        example: 1
        type: integer
      uuid:
        example: 998f91f3-4dd7-419d-a543-0d26a0e945ec
        type: string
    type: object
  admin_product_handler.CreateStockRequestDto:
    properties:
      minQty:
        example: 50
        type: integer
      qty:
        example: 100
        minimum: 1
        type: integer
    required:
    - qty
    type: object
...
  admin_product_handler.GetAllResponseDto:
    properties:
      active:
        example: true
        type: boolean
      categoryId:
        example: 3
        type: integer
      createdAt:
        example: "2023-01-01T00:00:00Z"
        type: string
      createdBy:
        example: 1
        type: integer
      description:
        example: Fresh and fiery red hot chilly peppers, perfect for adding a spicy
          kick to your dishes.
        type: string
      discountPrice:
        example: 4.99
        type: number
      id:
        example: 1
        type: integer
      image:
        example: https://example.com/images/red-hot-chilly-peppers.jpg
        type: string
      imageMobile:
        example: https://example.com/images/red-hot-chilly-peppers-mobile.jpg
        type: string
      imageThumbnail:
        example: https://example.com/images/red-hot-chilly-peppers-thumbnail.jpg
        type: string
      isDeleted:
        example: false
        type: boolean
      name:
        example: Red Hot Chilly Peppers
        type: string
      price:
        example: 5.99
        type: number
      updatedAt:
        example: "2023-01-01T00:00:00Z"
        type: string
      updatedBy:
        example: 1
        type: integer
      uuid:
        example: 998f91f3-4dd7-419d-a543-0d26a0e945ec
        type: string
    type: object
  admin_product_handler.GetAllWithRelationsResponseDto:
    properties:
      active:
        example: true
        type: boolean
      category:
        $ref: '#/definitions/admin_product_handler.CategoryResponseDto'
      categoryId:
        example: 3
        type: integer
      createdAt:
        example: "2023-01-01T00:00:00Z"
        type: string
      createdBy:
        example: 1
        type: integer
      description:
        example: Fresh and fiery red hot chilly peppers, perfect for adding a spicy
          kick to your dishes.
        type: string
      discountPrice:
        example: 4.99
        type: number
      id:
        example: 1
        type: integer
      image:
        example: https://example.com/images/red-hot-chilly-peppers.jpg
        type: string
      imageMobile:
        example: https://example.com/images/red-hot-chilly-peppers-mobile.jpg
        type: string
      imageThumbnail:
        example: https://example.com/images/red-hot-chilly-peppers-thumbnail.jpg
        type: string
      installments:
        $ref: '#/definitions/admin_product_handler.InstallmentsResponseDto'
      isDeleted:
        example: false
        type: boolean
      name:
        example: Red Hot Chilly Peppers
        type: string
      price:
        example: 5.99
        type: number
      stock:
        $ref: '#/definitions/admin_product_handler.StockResponseDto'
      updatedAt:
        example: "2023-01-01T00:00:00Z"
        type: string
      updatedBy:
        example: 1
        type: integer
      uuid:
        example: 998f91f3-4dd7-419d-a543-0d26a0e945ec
        type: string
    type: object
  admin_product_handler.GetOneByIdResponseDto:
    properties:
      active:
        example: true
        type: boolean
      category:
        $ref: '#/definitions/admin_product_handler.CategoryResponseDto'
      categoryId:
        example: 3
        type: integer
      createdAt:
        example: "2023-01-01T00:00:00Z"
        type: string
      createdBy:
        example: 1
        type: integer
      description:
        example: Fresh and fiery red hot chilly peppers, perfect for adding a spicy
          kick to your dishes.
        type: string
      discountPrice:
        example: 4.99
        type: number
      id:
        example: 1
        type: integer
      image:
        example: https://example.com/images/red-hot-chilly-peppers.jpg
        type: string
      imageMobile:
        example: https://example.com/images/red-hot-chilly-peppers-mobile.jpg
        type: string
      imageThumbnail:
        example: https://example.com/images/red-hot-chilly-peppers-thumbnail.jpg
        type: string
      isDeleted:
        example: false
        type: boolean
      name:
        example: Red Hot Chilly Peppers
        type: string
      price:
        example: 5.99
        type: number
      stock:
        $ref: '#/definitions/admin_product_handler.StockResponseDto'
      updatedAt:
        example: "2023-01-01T00:00:00Z"
        type: string
      updatedBy:
        example: 1
        type: integer
      uuid:
        example: 998f91f3-4dd7-419d-a543-0d26a0e945ec
        type: string
    type: object
  admin_product_handler.GetOneByUuidResponseDto:
    properties:
      active:
        example: true
        type: boolean
      category:
        $ref: '#/definitions/admin_product_handler.CategoryResponseDto'
      categoryId:
        example: 3
        type: integer
      createdAt:
        example: "2023-01-01T00:00:00Z"
        type: string
      createdBy:
        example: 1
        type: integer
      description:
        example: Fresh and fiery red hot chilly peppers, perfect for adding a spicy
          kick to your dishes.
        type: string
      discountPrice:
        example: 4.99
        type: number
      id:
        example: 1
        type: integer
      image:
        example: https://example.com/images/red-hot-chilly-peppers.jpg
        type: string
      imageMobile:
        example: https://example.com/images/red-hot-chilly-peppers-mobile.jpg
        type: string
      imageThumbnail:
        example: https://example.com/images/red-hot-chilly-peppers-thumbnail.jpg
        type: string
      isDeleted:
        example: false
        type: boolean
      name:
        example: Red Hot Chilly Peppers
        type: string
      price:
        example: 5.99
        type: number
      stock:
        $ref: '#/definitions/admin_product_handler.StockResponseDto'
      updatedAt:
        example: "2023-01-01T00:00:00Z"
        type: string
      updatedBy:
        example: 1
        type: integer
      uuid:
        example: 998f91f3-4dd7-419d-a543-0d26a0e945ec
        type: string
    type: object
...
  admin_product_handler.UpdateRequestDto:
    properties:
      active:
        example: false
        type: boolean
      categoryId:
        example: 1
        minimum: 1
        type: integer
      description:
        example: Update Fresh and fiery red hot chilly peppers, perfect for adding
          a spicy kick to your dishes.
        maxLength: 2000
        type: string
      discountPrice:
        example: 5.11
        type: number
      image:
        example: https://example.com/images/red-hot-chilly-peppers.png
        maxLength: 2000
        type: string
      imageMobile:
        example: https://example.com/images/red-hot-chilly-peppers-mobile.png
        maxLength: 2000
        type: string
      imageThumbnail:
        example: https://example.com/images/red-hot-chilly-peppers-thumbnail.png
        maxLength: 2000
        type: string
      installments:
        items:
          $ref: '#/definitions/admin_product_handler.UpdateInstallmentsRequestDto'
        minItems: 1
        type: array
      name:
        example: Update Red Hot Chilly Peppers
        maxLength: 255
        type: string
      price:
        example: 6.11
        type: number
    required:
    - categoryId
    - image
    - imageMobile
    - imageThumbnail
    - installments
    - name
    - price
    type: object
...
  creditcard_handler.CreateRequestDto:
    properties:
      creditcardFlagId:
        example: 1
        minimum: 1
        type: integer
      creditcardTypeId:
        example: 1
        minimum: 1
        type: integer
      expiration:
        example: "2025-01-01"
        type: string
      name:
        example: Testing
        type: string
      notifyExpiration:
        example: true
        type: boolean
      number:
        example: "4242424242424242"
        minLength: 8
        type: string
      priority:
        example: 1
        type: integer
      tokenize:
        example: true
        type: boolean
    required:
    - creditcardFlagId
    - creditcardTypeId
    - expiration
    - name
    - number
    - tokenize
    type: object
  creditcard_handler.CreateResponseDto:
    properties:
      createdAt:
        example: "2024-01-01T00:00:00Z"
        type: string
      creditcardFlagId:
        example: 1
        minimum: 1
        type: integer
      creditcardTypeId:
        example: 1
        minimum: 1
        type: integer
      expiration:
        example: "2025-01-01"
        type: string
      id:
        example: 1
        type: integer
      isDeleted:
        example: false
        type: boolean
      name:
        example: Testing
        type: string
      notifyExpiration:
        example: true
        type: boolean
      number:
        example: 4242********4242
        minLength: 8
        type: string
      priority:
        example: 1
        type: integer
      updatedAt:
        example: "2024-01-01T00:00:00Z"
        type: string
      userId:
        example: 1
        type: integer
      uuid:
        example: 767c1587-1ade-4e6a-a9fe-f0e07c83bda9
        type: string
    required:
    - creditcardFlagId
    - creditcardTypeId
    - expiration
    - name
    - number
    type: object
...
  creditcard_handler.GetAllUserCreditcardsResponseDto:
    properties:
      createdAt:
        example: "2024-01-01T00:00:00Z"
        type: string
      creditcardFlag:
        $ref: '#/definitions/creditcard_handler.CreditcardFlag'
      creditcardType:
        $ref: '#/definitions/creditcard_handler.CreditcardType'
      expiration:
        example: "2025-01-01T00:00:00Z"
        type: string
      name:
        example: Testing
        type: string
      number:
        example: 4242********4242
        minLength: 8
        type: string
      userId:
        example: 1
        type: integer
      uuid:
        example: 767c1587-1ade-4e6a-a9fe-f0e07c83bda9
        type: string
    required:
    - expiration
    - name
    - number
    type: object
  creditcard_handler.UpdateRequestDto:
    properties:
      creditcardFlagId:
        example: 2
        minimum: 1
        type: integer
      creditcardTypeId:
        example: 2
        minimum: 1
        type: integer
      expiration:
        example: "2025-01-01"
        type: string
      name:
        example: Testing Update
        type: string
      notifyExpiration:
        example: true
        type: boolean
      number:
        example: "6643213384289379"
        minLength: 8
        type: string
      priority:
        example: 2
        type: integer
    required:
    - creditcardFlagId
    - creditcardTypeId
    - expiration
    - name
    - number
    type: object
  exception.HttpException:
    properties:
      message:
        example: Invalid path parameter
        type: string
      statusCode:
        example: 400
        type: integer
    type: object
  middleware.AuthenticationError:
    properties:
      message:
        example: Access denied
        type: string
      statusCode:
        example: 401
        type: integer
    type: object
  middleware.RolePermissionError:
    properties:
      message:
        example: permission denied.
        type: string
      statusCode:
        example: 403
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: Ticket Sale Support
    url: http://www.swagger.io/support
  description: This is a simple ticket sales application.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Ticket Sale
  version: "1.0"
paths:
...
  /admin/products:
    get:
      description: Get all products
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/admin_product_handler.GetAllResponseDto'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.HttpException'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.AuthenticationError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middleware.RolePermissionError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.HttpException'
      summary: Get all products
      tags:
      - Admin Product
    post:
      description: Create a product
      parameters:
      - description: New product
        in: body
        name: Product
        required: true
        schema:
          $ref: '#/definitions/admin_product_handler.CreateRequestDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/admin_product_handler.CreateResponseDto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.HttpException'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.AuthenticationError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middleware.RolePermissionError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.HttpException'
      summary: Create a product
      tags:
      - Admin Product
  /admin/products/{id}:
    delete:
      description: Soft Delete a product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.HttpException'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.AuthenticationError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middleware.RolePermissionError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.HttpException'
      summary: Soft Delete a product
      tags:
      - Admin Product
    get:
      description: Get One By Id
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin_product_handler.GetOneByIdResponseDto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.HttpException'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.AuthenticationError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middleware.RolePermissionError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.HttpException'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.HttpException'
      summary: Get One By Id
      tags:
      - Admin Product
    put:
      description: Update a product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update product
        in: body
        name: Product
        required: true
        schema:
          $ref: '#/definitions/admin_product_handler.UpdateRequestDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.HttpException'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.AuthenticationError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middleware.RolePermissionError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.HttpException'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.HttpException'
      summary: Update a product
      tags:
      - Admin Product
  /admin/products/details:
    get:
      description: Get all products with relations
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/admin_product_handler.GetAllWithRelationsResponseDto'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.HttpException'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.AuthenticationError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middleware.RolePermissionError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.HttpException'
      summary: Get all products with relations
      tags:
      - Admin Product
  /admin/products/uuid/{uuid}:
    get:
      description: Get One By UUID
      parameters:
      - description: Product UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin_product_handler.GetOneByUuidResponseDto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.HttpException'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.AuthenticationError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middleware.RolePermissionError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.HttpException'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.HttpException'
      summary: Get One By UUID
      tags:
      - Admin Product
...
  /creditcard:
    post:
      description: Create a Creditcard
      parameters:
      - description: New Creditcard
        in: body
        name: Creditcard
        required: true
        schema:
          $ref: '#/definitions/creditcard_handler.CreateRequestDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/creditcard_handler.CreateResponseDto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.HttpException'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.AuthenticationError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middleware.RolePermissionError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.HttpException'
      summary: Create a Creditcard
      tags:
      - Creditcard
  /creditcard/{uuid}:
    delete:
      description: Soft Delete a creditcard
      parameters:
      - description: Creditcard UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.HttpException'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.AuthenticationError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middleware.RolePermissionError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.HttpException'
      summary: Soft Delete a creditcard
      tags:
      - Creditcard
    put:
      description: Update a creditcard
      parameters:
      - description: Creditcard UUID
        in: path
        name: uuid
        required: true
        type: string
      - description: Update Creditcard
        in: body
        name: Creditcard
        required: true
        schema:
          $ref: '#/definitions/creditcard_handler.UpdateRequestDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.HttpException'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.AuthenticationError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middleware.RolePermissionError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.HttpException'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.HttpException'
      summary: Update a creditcard
      tags:
      - Creditcard
  /creditcard/user:
    get:
      description: Get all user creditcards
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/creditcard_handler.GetAllUserCreditcardsResponseDto'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.HttpException'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middleware.AuthenticationError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/middleware.RolePermissionError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.HttpException'
      summary: Get all user creditcards
      tags:
      - Creditcard
...
swagger: "2.0"